restructure: lightning get api entities 폴더로 이동
