+++ src/entities/meeting/api/fetchMeetings.ts
