+++ src/shared/hooks/useLightningMeetups.ts
@@ -1,5 +1,5 @@
 import { useQuery, keepPreviousData } from '@tanstack/react-query';
-import { fetchLightningMeetups } from '@/shared/api/fetchMeetups';
+import { fetchLightningMeetups } from '@/entities/lightning/api/fetchMeetups';
 import { LightningMeetup } from '@/shared/types/lightningType';
 
 interface Filters {
